backend/requirements.txt
"fastapi[standard]"
numpy
python-whois
transformers
google-genai
scikit-learn
pyzbar
torch
pytesseract
opencv-python
pillow